Add independant settings menu
